requirements.txt
mcp>=1.23.0
pyautogui>=0.9.54
pillow>=12.0.0

